Feature/swap

+++ app/src/main/java/com/p2p/wallet/main/model/TokenPrice.kt
@@ -7,5 +7,5 @@ import java.math.BigDecimal
 @Parcelize
 data class TokenPrice(
     val tokenSymbol: String,
-    val price: BigDecimal
+    val price: Double
 ) : Parcelable
